requirements.txt
Flask==2.1.1
Flask_RESTful==0.3.9
Flask_Table==0.5.0
generators==2020.4.27
pandas
psycopg2_binary==2.9.3
pyspark==3.2.1
pytest==7.1.1
requests==2.27.1
WTForms==3.0.1
